+++ census/views/macros/popovers.html
@@ -19,13 +19,8 @@
   <ul class="availability-popover">
     {% for question in questions -%}
       {% if question.score -%}
-        {% if is_index %}
-          {% set answer = entry.answers|find({'id': question.id}) %}
-          {% set pass = 'yes' if answer.pass else 'no' %}
-        {% else %}
-          {% set answer = entry.getAnswerValueForQuestion(question) %}
-          {% set pass = 'yes' if question.pass(answer) else 'no' %}
-        {% endif %}
+        {% set answer = entry.getAnswerValueForQuestion(question) %}
+        {% set pass = 'yes' if question.pass(answer) else 'no' %}
         <li class="{{ pass }}" title="{{ question.question }}">
           <i class="fa fa-{{ question.icon }}"></i>&nbsp;{% if pass == 'yes' %}It's {% elif pass == 'no' %}It's not {% endif %}{{ question.questionshort|lower }}
         </li>
